refactor: move launchactivity into flow package
