+++ app2/src/main/java/com/chesire/malime/injection/androidmodules/ActivityModule.kt
@@ -1,6 +1,6 @@
 package com.chesire.malime.injection.androidmodules
 
-import com.chesire.malime.LaunchActivity
+import com.chesire.malime.flow.LaunchActivity
 import dagger.Module
 import dagger.android.ContributesAndroidInjector
